реализовано удаление клиентов, рабочий сервис

+++ GateWay/src/main/java/ru/sem/gateway/client/ClientDto.java
@@ -23,7 +23,7 @@ public class ClientDto {
     @NotBlank
     String nickName;
 
-    Integer phone;
+    Long phone;
 
     @Email
     String email;
